[FIX/#81] 수정된 api 명세서에 맞게 controller를 수정해라
