+++ api/src/main/java/com/org/gunbbang/config/SecurityConfig.java
@@ -58,8 +58,8 @@ public class SecurityConfig {
                 // 아이콘, css, js 관련
                 // 기본 페이지, css, image, js 하위 폴더에 있는 자료들은 모두 접근 가능, h2-console에 접근 가능
                 .antMatchers("/","/css/**","/images/**","/js/**","/favicon.ico","/h2-console/**", "/health", "/profile").permitAll()
-                .antMatchers("/member/signup", "/validate/nickname", "/validate/email").permitAll() // 회원가입, 닉네임 중복체크, 이메일 중복체크는은모든 리소스에 접근 가능
-                .antMatchers("/health", "/profile", "/actuator/health").permitAll()
+                .antMatchers("/auth/signup", "/validation/nickname", "/validation/email").permitAll() // 회원가입, 닉네임 중복체크, 이메일 중복체크는은모든 리소스에 접근 가능
+                .antMatchers("/profile", "/actuator/health").permitAll()
                 .anyRequest().authenticated();
 
         // 원래 스프링 시큐리티 필터 순서가 LogoutFilter 이후에 로그인 필터 동작
