+++ api/src/main/java/com/org/gunbbang/controller/BakeriesController.java
@@ -1,9 +1,12 @@
 package com.org.gunbbang.controller;
 
 import com.org.gunbbang.common.dto.ApiResponse;
+import com.org.gunbbang.controller.DTO.response.BakeryDetailResponseDTO;
 import com.org.gunbbang.controller.DTO.response.BakeryListResponseDTO;
+import com.org.gunbbang.controller.DTO.response.ReviewListResponseDTO;
 import com.org.gunbbang.errorType.SuccessType;
 import com.org.gunbbang.service.BakeryService;
+import com.org.gunbbang.service.ReviewService;
 import com.org.gunbbang.util.Security.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -18,6 +21,8 @@ import java.util.List;
 public class BakeriesController {
 
     private final BakeryService bakeryService;
+    private final ReviewService reviewService;
+
     @GetMapping("")
     @ResponseStatus(HttpStatus.OK)
     public ApiResponse<List<BakeryListResponseDTO>> getBakeryList(@RequestParam("sort") @Valid String sort,
@@ -28,10 +33,18 @@ public class BakeriesController {
         List<BakeryListResponseDTO> bakeryListResponseDto = bakeryService.getBakeryList(memberId, sort,isHard,isDessert,isBrunch);
         return ApiResponse.success(SuccessType.GET_BAKERY_LIST_SUCCESS, bakeryListResponseDto);
     }
-
-    @GetMapping("/bookMark")
-    public ApiResponse<List<BakeryListResponseDTO>> getBookMarkedBakeries() {
+    @GetMapping("/{bakeryId}")
+    @ResponseStatus(HttpStatus.OK)
+    public ApiResponse<BakeryDetailResponseDTO> getBakeryDetail(@PathVariable("bakeryId") @Valid Long bakeryId){
         Long memberId = SecurityUtil.getLoginMemberId();
-        return ApiResponse.success(SuccessType.GET_BOOKMARKED_BAKERIES, bakeryService.getBookMarkedBakeries(memberId));
+        BakeryDetailResponseDTO bakeryDetailResponseDTO = bakeryService.getBakeryDetail(memberId, bakeryId);
+        return ApiResponse.success(SuccessType.GET_BAKERY_DETAIL_SUCCESS, bakeryDetailResponseDTO);
+    }
+
+    @GetMapping("/{bakeryId}/reviews")
+    @ResponseStatus(HttpStatus.OK)
+    public ApiResponse<ReviewListResponseDTO> getBakeryReviewList(@PathVariable("bakeryId") Long bakeryId) {
+        ReviewListResponseDTO reviewListResponseDto= reviewService.getBakeryReviewList(bakeryId);
+        return ApiResponse.success(SuccessType.GET_BAKERY_REVIEW_LIST_SUCCESS, reviewListResponseDto);
     }
 }
