+++ api/src/main/java/com/org/gunbbang/controller/MemberController.java
@@ -2,11 +2,9 @@ package com.org.gunbbang.controller;
 
 import com.org.gunbbang.common.dto.ApiResponse;
 import com.org.gunbbang.controller.DTO.request.MemberTypesRequestDTO;
-import com.org.gunbbang.controller.DTO.response.BakeryListReviewedByMemberDTO;
-import com.org.gunbbang.controller.DTO.response.MemberDetailResponseDTO;
-import com.org.gunbbang.controller.DTO.response.MemberSignUpResponseDTO;
-import com.org.gunbbang.controller.DTO.response.MemberTypesResponseDTO;
+import com.org.gunbbang.controller.DTO.response.*;
 import com.org.gunbbang.errorType.SuccessType;
+import com.org.gunbbang.service.BakeryService;
 import com.org.gunbbang.service.MemberService;
 import com.org.gunbbang.service.ReviewService;
 import lombok.RequiredArgsConstructor;
@@ -16,6 +14,7 @@ import org.springframework.web.bind.annotation.*;
 import com.org.gunbbang.controller.DTO.request.MemberSignUpRequestDTO;
 import com.org.gunbbang.util.Security.SecurityUtil;
 
+import javax.validation.Valid;
 import java.util.List;
 
 
@@ -26,17 +25,13 @@ public class MemberController {
 
     private final MemberService memberService;
     private final ReviewService reviewService;
+    private final BakeryService bakeryService;
 
     @GetMapping("")
     @ResponseStatus(HttpStatus.OK)
     public ApiResponse<MemberDetailResponseDTO> getMemberDetail() {
         MemberDetailResponseDTO memberDetailResponseDto = memberService.getMemberDetail();
         return ApiResponse.success(SuccessType.GET_MYPAGE_SUCCESS, memberDetailResponseDto);
-    }
-
-    @PostMapping("/signup")
-    public ApiResponse<MemberSignUpResponseDTO> signUp(@RequestBody final MemberSignUpRequestDTO memberSignUpRequestDTO) throws Exception {
-        return ApiResponse.success(SuccessType.SIGNUP_SUCCESS, memberService.signUp(memberSignUpRequestDTO));
     }
 
     @GetMapping("/security-test")
@@ -46,7 +41,7 @@ public class MemberController {
 
     @PostMapping("/types")
     @ResponseStatus(HttpStatus.OK)
-    public ApiResponse<MemberTypesResponseDTO> updateMemberTypes(@RequestBody final MemberTypesRequestDTO request) {
+    public ApiResponse<MemberTypesResponseDTO> updateMemberTypes(@Valid @RequestBody final MemberTypesRequestDTO request) {
         Long memberId = SecurityUtil.getLoginMemberId();
         return ApiResponse.success(SuccessType.UPDATE_MEMBER_TYPES_SUCCESS, memberService.updateMemberTypes(request, memberId));
     }
@@ -63,5 +58,11 @@ public class MemberController {
     public ApiResponse<List<BakeryListReviewedByMemberDTO>> getBakeryListReviewedByMember(){
         Long memberId = SecurityUtil.getLoginMemberId();
         return ApiResponse.success(SuccessType.GET_MEMBER_REVIEW_BAKERY_LIST_SUCCESS, reviewService.getBakeryListReviewedByMember(memberId));
+    }
+
+    @GetMapping("/bookMarks")
+    public ApiResponse<List<BakeryListResponseDTO>> getBookMarkedBakeries() {
+        Long memberId = SecurityUtil.getLoginMemberId();
+        return ApiResponse.success(SuccessType.GET_BOOKMARKED_BAKERIES, bakeryService.getBookMarkedBakeries(memberId));
     }
 }
